Edit tasks added

+++ ChadCalendar/ChadCalendar/Models/Task.cs
@@ -13,7 +13,7 @@ namespace ChadCalendar.Models
         [Required]
         public bool AllowedToDistribute { get; set; } = false; // избежание null
         [Required]
-        public decimal? HoursTakes { get; set; }
+        public int? HoursTakes { get; set; }
         public int? MaxPerDay { get; set; }
         public DateTime? Deadline { get; set; } = DateTime.Now.AddDays(1);
         public int? PredecessorFK { get; set; }
